ENGR00243299 Fixed SystemUI crash when disable GPU acceleration
The designated logic should switch to SW Canvas render if OpenGL-ES 2.0
is not available when GPU 3D core is disabled or in-existed.

The fix is for debug purpose on i.mx6d/q devices when disable GPU 3D.
Also enable JB4.2 on i.mx6sl device without GPU 3D core.

Change-Id: Iaaca554c9d5115ebf91426fe02b241a920ec8ee6
Signed-off-by: Xianzhong <b07117@freescale.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/306
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/927